% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform K Means Clustering.R
\name{perform_kmeans}
\alias{perform_kmeans}
\title{Perform K-Means Clustering}
\usage{
perform_kmeans(data, x, y, k = 3)
}
\arguments{
\item{data}{The dataset to perform clustering on (data frame).}

\item{x}{The x variable for clustering (a column name in the dataset).}

\item{y}{The y variable for clustering (a column name in the dataset).}

\item{k}{The number of clusters (default is 3).}
}
\value{
A list with cluster assignments and cluster centers.
}
\description{
This function performs K-Means clustering on two specified columns from the olive dataset.
}
