% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot K Means Clustering.R
\name{plot_kmeans}
\alias{plot_kmeans}
\title{Plot K-Means Clustering Results}
\usage{
plot_kmeans(data, kmeans_result, x, y)
}
\arguments{
\item{data}{The dataset used for clustering.}

\item{kmeans_result}{The result from \code{perform_kmeans}.}

\item{x}{The x variable for clustering (a column name in the dataset).}

\item{y}{The y variable for clustering (a column name in the dataset).}
}
\value{
A scatter plot showing the K-Means clusters.
}
\description{
This function will plot the outcome of K-Means clustering.
}
